cfxstore

Function random_phrase

pub fn random_phrase(no_of_words: usize) -> String

Generate a string which is a random phrase of a number of lowercase words.

words is the number of words, chosen from a dictionary of 7,530. An value of 12 gives 155 bits of entropy (almost saturating address space); 20 gives 258 bits which is enough to saturate 32-byte key space