cfxstore

Struct CfxMultiStore

pub struct CfxMultiStore { /* private fields */ }

Similar to CfxStore but may store many accounts (with different passwords) for the same Address

Implementations

impl CfxMultiStore

pub fn open(directory: Box<dyn KeyDirectory>) -> Result<Self, Error>

Open new multi-accounts store with given key directory backend.

pub fn open_with_iterations( directory: Box<dyn KeyDirectory>, iterations: u32, ) -> Result<Self, Error>

Open new multi-accounts store with given key directory backend and custom number of iterations for new keys.

pub fn set_refresh_time(&self, time: Duration)

Modify account refresh timeout - how often they are re-read from KeyDirectory.

Setting this to low values (or 0) will cause new accounts to be picked up quickly, although it may induce heavy disk reads and is not recommended if you manage many keys (say over 10k).

By default refreshing is disabled, so only accounts created using this instance of CfxStore are taken into account.