cfxstore

Module cfxkey

cfxkey reexport to make documentation look pretty.

Modules

brain_recover

crypto

math

Structs

Brain

Simple brainwallet.

BrainPrefix

Tries to find brain-seed keypair with address starting with given prefix.

ExtendedKeyPair

ExtendedPublic

Extended public key, allows deterministic derivation of subsequent keys.

ExtendedSecret

Extended secret key, allows deterministic derivation of subsequent keys.

KeyPair

secp256k1 key pair

Password

Prefix

Tries to find keypair with address starting with given prefix.

Random

Randomly generates new keypair, instantiating the RNG each time.

SECP256K1

Secret

Signature

Signature encoded as RSV components

Enums

Derivation

Key derivation over generic label T

DerivationError

Error

Crypto error

Void

Uninstantiatable error type for infallible generators.

WordlistError

Phrase Validation Error

Traits

Generator

Generates new keypair.

KeyPairGenerator

Generates new keypair.

Functions

is_compatible_public

Check if the recovered address started with 0x1. If it does, this public key will have same the address in Conflux Network and Ethereum.

public_is_valid

Whether the public key is valid.

public_to_address

recover

sign

verify_address

verify_public

Type Aliases

Address

Message

Public